sevenSegDecoder done
the 7-seg function is done, I created the segments in the abc_defg format because it was easier to create it. Dr Harris made us do things a bit backwards when I look back on past labs, so when we create the wrapper, we have to line the bits up backwards. For example, if we want to light up HEX3, we do HEX[0:6] instead of HEX[6:0]